添加了收藏页面碎片，但还没写布局和逻辑代码。

+++ app/src/main/java/com/hnqcgc/redfirecookbook/logic/Repository.java
@@ -82,7 +82,7 @@ public class Repository {
         return RecipeCountDao.getInstance().getRecipeCount();
     }
 
-    public int insertCollection(Collection collection) {
+    public long insertCollection(Collection collection) {
         return RedFireCookBookDB.getInstance().insertCollection(collection);
     }
